Login page design finalization
